+++ web/listAccount.jsp
@@ -134,7 +134,7 @@
                                                 <td class="text-right font-weight-semibold align-middle p-4">
                                                     <button><a><a href="updateAccount?updateId=${C.id}">Update</a></button>
                                                 </td>
-                                                <td class="text-center align-middle px-0"><a href="deleteAccount?deleteId=${C.id}" class="shop-tooltip close float-none text-danger" title="" data-original-title="Remove">×</a></td>
+                                                <td class="text-center align-middle px-0"><a href="deleteAccount?deleteId=${C.id}" class="shop-tooltip close float-none text-danger " onclick="return confirm('Bạn có đồng ý xoá tài khoản này?') title="" data-original-title="Remove">×</a></td>
                                             </tr>
                                         </c:forEach>
                                     </tbody>
